+++ Jarvis.Framework.Shared/JarvisFrameworkGlobalConfiguration.cs
@@ -5,7 +5,11 @@ namespace Jarvis.Framework.Shared
 {
     public static class JarvisFrameworkGlobalConfiguration
     {
-        public static Boolean IsMongodbLinq3Enabled { get; private set; }
+        /// <summary>
+        /// From version 7.8.2 Linq3 is enabled by default. You can disable if you want
+        /// but remember that version 2 will be removed in future version of Mongodb driver.
+        /// </summary>
+        public static Boolean IsMongodbLinq3Enabled { get; private set; } = true;
 
         public static Boolean OfflineEventsReadmodelIdempotencyCheck { get; private set; }
 
@@ -32,11 +36,13 @@ namespace Jarvis.Framework.Shared
         }
 
         /// <summary>
-        /// Enable mongodb driver linq3 support (introduced with 2.19 version of driver and full of bug in that version);
+        /// Enable mongodb driver linq2, if you have problem with the new default that is version 3 of Mongodb Linq provider.
+        /// Version 3 historically has bug, but now seems to be stable and version 2 will be removed in future version so
+        /// evaluate if you really need to use version 2 as fallback, because will be removed in future versions.
         /// </summary>
-        public static void EnableMongodbDriverLinq3()
+        public static void EnableMongodbDriverLinq2()
         {
-            IsMongodbLinq3Enabled = true;
+            IsMongodbLinq3Enabled = false;
         }
 
         public static void EnableOfflineEventsReadmodelIdempotencyCheck()