+++ Wiki/ReleaseNotes.md
@@ -1,5 +1,10 @@
 # J.A.R.V.I.S. Framework - Proximo srl (c)
 
+## 7.8.2
+
+- Linq driver for MongoDB now default to version 3.
+- Better support for cancellation token. 
+
 ## 7.8.1
 
 - Updated nstore, mongodb, and test libraries.